GH-28 - AggregateRoots now implement Persistable in JPA.
Types implementing AggregateRoot now automatically implement Persistable. They carry a synthetic is-new field that is exposed via Persistable.isNew() and flipped in @PrePersist / @PostConstruct callbacks. Also, we implement equals(…) and hashCode() based on the singular identifier typed field of the entity.

Note, due to issues with augmenting the nullability verification method [0], we've temporarily disabled the generation of that method.

[0] https://stackoverflow.com/questions/66489614/how-to-augment-methods-during-bytebuddy-transformation