GH-74 - Improve code generation for nullability verification method.
We now make sure that we do not generate multiple entity lifecycle callback methods in case the user code already contains one. If that's the case, we rather advice the existing method to get the null to verify the nullability on method exit, so that callbacks could still default some of the fields declared as non-nullable.