GH-85 - Ensure Jakarta Persistence 3.0 compatibility.
We now make sure that the JPA-related code generation applies to both JPA (or Jakarta Persistence) up to 2.2 code is generated just as for Jakarta Persistence 3.0.

Introduced Jpa as an abstraction to obtain JPA / Jakarta Persistence annotations basically just changing the base package to refer to the types to be loaded. The individual Plugin implementations now all also implement WithPreprocessor via JMoleculesPluginSupport to be able to initialize a Jpa instance when they're used declaratively, as they are in the tests.