GH-92, GH-93 - Improve JPA default mappings.
We now also trigger JPA relationship defaulting for annotated entities and aggregates. Tweaked the default identifier mapping to use @EmbeddedId if the field's type implements Identifier. Resort back to @Id in case a simple @Identity field has been found.

Tweaked the defaults for the to-many relationships to only use eager mapping if there's exactly one. In case of multiple to-many relationships, we now explicitly configure a lazy relationship. In case Hibernate is used, lazy relationships get augmented with an @Fetch(FetchMode.SUBSELECT) to make sure the entire collection is materialized in one go, which we can do because we can safely assume collection relationships to entities are never unbounded.

We now also default to-entity relationships to use orphan removal so that entities removed from the collection get removed from the database automatically.