GH-98 - Support for transparent JPA persistence of Java records as embeddables on Hibernate 6.
We now provide a Hibernate 6 EmbeddableInstantiator implementation that will instantiate Java records so that can be transparently used as embeddables in domain models persisted with JPA. The class can be used explicitly by declaring a subtype invoking the super class' constructor with the record type.

The jMolecules ByteBuddy plugin was extended to automatically generate that class for every record type it encounters also annotating the type with @Embeddable and @EmbeddableInstantiator to avoid the additional noise on the domain code.

Had to tweak the build to use Java 17 in tests to be able to actually declare records in the first place. Also, although we still retain Java 8 compatibility, we need to run the build with a JDK 15+ now to be able to see the records related reflection API to analyze the record's components.