GH-110 - Make ArchUnit DDD rules more lenient against non matching rules.
ArchUnit fails if no source code elements match the original selector to apply rules against. This is primarily to detect potential selector mismatches that result in no violations found. In our case, we don't want to fail for those as it just might mean that not source elements that need to match exist.