GH-133 - Add no-dependency flavor of ByteBuddy plugin.
For use with the new user compile classpath ByteBuddy plugin detection, we now ship a shaded variant of our ByteBuddy plugin that does not have any dependencies declared. We need the dependencies in the original JAR for usage in the traditional way as only this makes sure that the plugin execution sees the JARs. However, for usage with the new mechanism we need to reduce the footprint of dependencies to make sure we don't pollute the user compile classpath.

We now also avoid a runtime dependency to Spring Core solely for assertions.