GH-180, GH-181 - Improvements in RecordInstantiator.
RecordInstantiator now works around a bug [0] in Hibernate 6.2 to make sure that record component values are handed into a record's constructor in the correct order.

Removed all references to Spring from RecordInstantiator to be able to move it into the jmolecules-jpa artifact. Adapted ByteBuddy code generation to move the application of the instantiator into the JMoleculesJpaPlugin.

[0] https://hibernate.atlassian.net/browse/HHH-16457