GH-200 - Support for dedicated stereotype lookups in ArchUnit verifications.
We now support to customize the lookup of usually package-level annotations on a marker type so that the definition would be extended to the entire package the marker type resides in. This can be considered an alternative to annotating package-info.java in, for example, languages that do not support package annotations, such as Kotlin.