GH-207 - Fix callback generation for aggregates.
In case of annotation based callbacks to flip the is new state of an aggregate we currently implement MutablePersistable but do not actually generate an implementation for the method declared by that interface. This doesn't seem to cause any problems currently as nobody is ever invoking that method (as the flipping is implemented through the annotation-based callbacks).

We now avoid implementing MutablePersistable for on-entity callbacks and fall back to only implementing Persistable, a getId() and isNew() method (the latter already generated before).