GH-230 - Guard APT against multiple processor invocations.
Multiple runs of the verifying APT process previously failed (hence the disabled tests). This seems to have to do with the need to alway work with the current "type pool" (APT's Types). The static references to TypeMirror instances in JMoleculesDddVerification previously caused assignability checks for types to fail in subsequent runs of the processor as they ultimately referenced the same type represented by a different instance (from a different type pool).